refactored PostItems
