+++ components/Posts/TextPost.js
@@ -1,22 +1,18 @@
 import { useNavigation } from "@react-navigation/native";
-import { Text, View } from "react-native";
+import { View, Text } from "react-native";
 import { getAge } from "../../features/shared/utils/getAge";
 import { useParseContent } from "../../hooks/useParseContent";
 import { useZapNote } from "../../hooks/useZapNote";
-import globalStyles from "../../styles/globalStyles";
-import FeedImage from "../Images/FeedImage";
 import PostActionBar from "./PostActionBar";
 
-const ImagePost = ({ event, user, width }) => {
+const TextPost = ({ event, user, width }) => {
     const content = useParseContent(event);
     const navigation = useNavigation();
-
     const zap = useZapNote(
         event.id,
         user?.lud06 || user?.lud16,
         user?.name || event?.pubkey.slice(0, 16)
     );
-
     const commentHandler = () => {
         navigation.navigate("CommentScreen", {
             eventId: event.id,
@@ -35,20 +31,19 @@ const ImagePost = ({ event, user, width }) => {
     return (
         <View
             style={{
+                padding: 6,
+                borderRadius: 6,
                 marginBottom: 12,
             }}
         >
             <View
                 style={{
+                    width: "100%",
                     flexDirection: "row",
                     justifyContent: "space-between",
-                    width: "100%",
-                    alignItems: "center",
                     borderBottomColor: colors.primary500,
                     borderBottomWidth: 1,
                     paddingBottom: 6,
-                    marginBottom: 6,
-                    paddingHorizontal: 6,
                 }}
             >
                 <Text
@@ -60,28 +55,27 @@ const ImagePost = ({ event, user, width }) => {
                 >
                     {user?.name || event.pubkey}
                 </Text>
-                <Text
-                    style={[
-                        globalStyles.textBodyS,
-                        { textAlign: "right", marginTop: 12 },
-                    ]}
-                >
+                <Text style={[globalStyles.textBodyS]}>
                     {getAge(event.created_at)}
                 </Text>
             </View>
-            <FeedImage size={width - 12} images={event.image} />
-            <View style={{paddingHorizontal: 6}}>
-                <Text style={[globalStyles.textBody, { textAlign: "left" }]}>
-                    {content}
-                </Text>
-                <PostActionBar
-                    onPressComment={commentHandler}
-                    onPressZap={zapHandler}
-                    onPressMore={moreHandler}
-                />
-            </View>
+
+            <Text
+                style={[
+                    globalStyles.textBody,
+                    { textAlign: "left", marginTop: 16 },
+                ]}
+            >
+                {content}
+            </Text>
+            <PostActionBar
+                onPressComment={commentHandler}
+                onPressZap={zapHandler}
+                onPressMore={moreHandler}
+                zapDisabled={!user?.lud06 && !user?.lud16}
+            />
         </View>
     );
 };
 
-export default ImagePost;
+export default TextPost;
